+++ packages/queue/src/router/index.ts
@@ -13,6 +13,7 @@ export {
     isValidPattern,
     isValidTopic,
     matchTopic,
+    matchTopicParts,
     MULTI_WILDCARD,
     SINGLE_WILDCARD,
     TOPIC_SEPARATOR,